Move isDisplayablePropMedia to mediaHelpers
